Move jose tests into src
