+++ identity_jose/src/tests/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2020-2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod es256;
+mod hs256;
+mod rfc7515;
+mod rfc7517;
+mod rfc7638;
+mod rfc7797;
